requirements.txt
arxiv==2.1.0
requests==2.31.0
google-generativeai==0.7.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client==2.125.0
